requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coverage==4.5.1
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-JWT-Extended==3.13.1
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
PyJWT==1.4.2
pytest==3.10.0
pytest-cov==2.6.0
requests==2.20.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1